Fixed project for samples, compiling Debug SDL1 and Release SDL1. Will fix SDL2 soon. Fixed minor warnings and errors on SDL1 (key input map), still need fix joystick mapping. Some samples does not work correctly or at all: joystick, movie, bullet, network.

+++ samples/03_pointer_input/pointer_sample.cpp
@@ -85,8 +85,8 @@ void PointerSample::OnInputPointerRelease(const EventInputPointer *ev)
 		if (pObject)
 			vFrom = pObject->GetPosition();
 
-		vTo.setX(ev->GetX());
-		vTo.setY(ev->GetY());
+		vTo.setX(f32(ev->GetX()));
+		vTo.setY(f32(ev->GetY()));
 		vTo += pCamera->GetPosition();
 		fElapsed = 0.0f;
 	}
