+++ samples/14_renderer/renderer_sample.cpp
@@ -132,8 +132,8 @@ void RendererSample::OnInputPointerRelease(const EventInputPointer *ev)
 		if (pObject)
 			vFrom = pObject->GetPosition();
 
-		vTo.setX(ev->GetX());
-		vTo.setY(ev->GetY());
+		vTo.setX(f32(ev->GetX()));
+		vTo.setY(f32(ev->GetY()));
 		vTo += pCamera->GetPosition();
 		fElapsed = 0.0f;
 	}
